requirements.txt
APScheduler==3.6.3
botocore==1.20.70
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.2.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-firebase==1.2
python-telegram-bot==13.0
pytz==2020.4
requests==2.24.0
s3transfer==0.4.2
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
